pyproject.toml
[project]
name = "magicpin-ai-challenge"
version = "0.1.0"
description = "Vera bot — magicpin AI Challenge merchant assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot.py", "server.py", "state.py", "llm_client.py", "validator.py", "classifiers.py", "make_submission.py", "prompts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
